# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  dto.ResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: Laundry APps
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login for User
      parameters:
      - description: Data login user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: Login
      tags:
      - User
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"